Code improved and cleaned.

+++ TriviaProject/Timer.h
@@ -33,7 +33,7 @@ public:
             endTime = m_EndTime;
         }
 
-        return std::chrono::duration_cast<std::chrono::milliseconds>(endTime - m_StartTime).count();
+        return static_cast<double>(std::chrono::duration_cast<std::chrono::milliseconds>(endTime - m_StartTime).count());
     }
     inline void eraseTimers()
     {
